ref(common>packet>update-game-state): Changed gameState type `uint8_t` --> `uint64_t`

+++ common/packets/server/update_game_state.hpp
@@ -16,7 +16,7 @@ struct UpdateGameState
     static constexpr auto kFlag = rtnt::core::packet::Flag::kReliable;
     static constexpr rtnt::core::packet::Name kName = "UPDATE_GAME_STATE";
 
-    uint8_t gameState;  ///< The current state of the game.
+    uint64_t gameState;  ///< The current state of the game.
 
     template <typename Archive>
     void serialize(Archive& ar)
